add a new plain-text volume ui inside the main window
The goal here is code simplification by having it in the main window.  Using plain text is because VLC does that and it looks basically fine, though the specific format used here probably needs tweaking (it's illegible on some backgrounds).

Need to fix the volume control to be even multiples of 10% (currently it arbitrarily ends up as things like 29% instead).