use a single overlay window that always covers the whole area of the main window
This simplifies things, because keeping a single overlay window's position/size in sync is trivial.  And having it always there means the titlebar/controlbar can use their own NSTrackingArea, that autosizes with their NSView, rather than having one set of tracking areas in the main window for showing them, and another in the subwindows for hiding them.