move the volume indicator to the overlay window
Because it was always ugly having it as a child of the movie-view (rather than a sibling) and would potentially intefere with changing video backend.

Also, use a black translucent background, rather than shadow, because it's readable over a wider variety of videos.  I'd have liked some padding on the rectangle, but NSTextField doesn't have any built-insupport for that.