show the full-screen control-bar at the bottom of the screen.
Rather than the bottom of the video.  I broke this when switching to just one overlay window.

Also, make the title never show in full-screen mode (because it's not much use, and I don't want to have to think about how to make it get out of the way of the menubar).