use LAVPLayer rather than LAVPView
So we don't need to subclass it just to implement -acceptsFirstMouse:

When I first made the choice to use LAVPView they actually used completely different code for getting pixels on screen and LAVPView was the only one I could get working.  But now LAVPView is just a trivial wrapper of LAVPLayer anyway.