fix the time-indicator to alway be the right width for the text shown there
Prior to this it was generally too narrow on any movie over an hour, which meant it wrapped, so the current time was visible but the duration ended up hidden.

We don't want to just statically make it wider, because having a large gap next to it on shorter movies looks ugly.  And we don't want to continuously resize it for the current time because having the scrubber keep changing size second-by-second would look weird.